src/frontend/requirements.txt
streamlit
pymongo
redis
python-dotenv
requests
pandas
ollama